backend_python_fastAPI/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-generativeai
python-dotenv
python-multipart
chromadb
bleach
